+++ packages/wxt/e2e/tests/hooks.test.ts
@@ -4,6 +4,7 @@ import { WxtHooks } from '~/types';
 
 const hooks: WxtHooks = {
   ready: vi.fn(),
+  'prepare:types': vi.fn(),
   'build:before': vi.fn(),
   'build:done': vi.fn(),
   'build:manifestGenerated': vi.fn(),
@@ -41,6 +42,7 @@ describe('Hooks', () => {
 
     expectHooksToBeCalled({
       ready: true,
+      'prepare:types': true,
       'build:before': false,
       'build:done': false,
       'build:publicAssets': false,
@@ -60,6 +62,7 @@ describe('Hooks', () => {
 
     expectHooksToBeCalled({
       ready: true,
+      'prepare:types': true,
       'build:before': true,
       'build:done': true,
       'build:publicAssets': true,
@@ -79,6 +82,7 @@ describe('Hooks', () => {
 
     expectHooksToBeCalled({
       ready: true,
+      'prepare:types': true,
       'build:before': true,
       'build:done': true,
       'build:publicAssets': true,
@@ -104,6 +108,7 @@ describe('Hooks', () => {
 
     expectHooksToBeCalled({
       ready: true,
+      'prepare:types': true,
       'build:before': true,
       'build:done': true,
       'build:publicAssets': true,