+++ packages/wxt/src/builtin-modules/index.ts
@@ -1,0 +1,4 @@
+import { WxtModule } from '~/types';
+import unimport from './unimport';
+
+export const builtinModules: WxtModule<any>[] = [unimport];