+++ packages/wxt/src/core/builders/vite/index.ts
@@ -61,7 +61,6 @@ export async function createViteBuilder(
     config.plugins.push(
       wxtPlugins.download(wxtConfig),
       wxtPlugins.devHtmlPrerender(wxtConfig, server),
-      wxtPlugins.unimport(wxtConfig),
       wxtPlugins.resolveVirtualModules(wxtConfig),
       wxtPlugins.devServerGlobals(wxtConfig, server),
       wxtPlugins.tsconfigPaths(wxtConfig),
@@ -245,7 +244,10 @@ export async function createViteBuilder(
           const config = vite.mergeConfig(baseConfig, envConfig);
           const server = await vite.createServer(config);
           await server.pluginContainer.buildStart({});
-          const node = new ViteNodeServer(server);
+          const node = new ViteNodeServer(
+            // @ts-ignore: Some weird type error...
+            server,
+          );
           installSourcemapsSupport({
             getSourceMap: (source) => node.getSourceMap(source),
           });