+++ packages/wxt/src/core/builders/vite/plugins/index.ts
@@ -2,7 +2,6 @@ export * from './devHtmlPrerender';
 export * from './devServerGlobals';
 export * from './download';
 export * from './multipageMove';
-export * from './unimport';
 export * from './resolveVirtualModules';
 export * from './tsconfigPaths';
 export * from './noopBackground';