+++ packages/wxt/src/core/clean.ts
@@ -1,44 +1,65 @@
 import path from 'node:path';
 import glob from 'fast-glob';
 import fs from 'fs-extra';
-import { consola } from 'consola';
 import pc from 'picocolors';
+import { InlineConfig } from '~/types';
+import { registerWxt, wxt } from './wxt';
 
 /**
  * Remove generated/temp files from the directory.
+ *
+ * @param config Optional config that will override your `<root>/wxt.config.ts`.
+ *
+ * @example
+ * await clean();
+ */
+export async function clean(config?: InlineConfig): Promise<void>;
+/**
+ * Remove generated/temp files from the directory.
+ *
+ * @deprecated
  *
  * @param root The directory to look for generated/temp files in. Defaults to `process.cwd()`. Can be relative to `process.cwd()` or absolute.
  *
  * @example
  * await clean();
  */
-export async function clean(root = process.cwd()) {
-  consola.info('Cleaning Project');
+export async function clean(root?: string): Promise<void>;
+
+export async function clean(config?: string | InlineConfig) {
+  if (typeof config === 'string') {
+    config = { root: config };
+  }
+
+  await registerWxt('build', config);
+  wxt.logger.info('Cleaning Project');
+
+  const root = wxt.config.root;
 
   const tempDirs = [
     'node_modules/.vite',
     'node_modules/.cache',
     '**/.wxt',
-    '.output/*',
+    `${path.relative(root, wxt.config.outBaseDir)}/*`,
   ];
-  consola.debug('Looking for:', tempDirs.map(pc.cyan).join(', '));
+  wxt.logger.debug('Looking for:', tempDirs.map(pc.cyan).join(', '));
   const directories = await glob(tempDirs, {
-    cwd: path.resolve(root),
+    cwd: root,
     absolute: true,
     onlyDirectories: true,
     deep: 2,
   });
   if (directories.length === 0) {
-    consola.debug('No generated files found.');
+    wxt.logger.debug('No generated files found.');
     return;
   }
 
-  consola.debug(
+  wxt.logger.debug(
     'Found:',
     directories.map((dir) => pc.cyan(path.relative(root, dir))).join(', '),
   );
   for (const directory of directories) {
     await fs.rm(directory, { force: true, recursive: true });
-    consola.debug('Deleted ' + pc.cyan(path.relative(root, directory)));
+    wxt.logger.debug('Deleted ' + pc.cyan(path.relative(root, directory)));
   }
 }