+++ packages/wxt/src/core/utils/package.ts
@@ -20,7 +20,3 @@ export async function getPackageJson(): Promise<
     return {};
   }
 }
-
-export function isModuleInstalled(name: string) {
-  return import(/* @vite-ignore */ name).then(() => true).catch(() => false);
-}